refactor : 검증 메소드 분리
