+++ src/main/java/onboarding/Problem1.java
@@ -35,21 +35,4 @@ class Problem1 {
         return answer;
     }
 
-    public static List<Integer> numPolar(List<Integer> pobi, List<Integer> crong) {
-        List<Integer> pages = new ArrayList<>(pobi);
-        pages.addAll(crong);
-        if (pages.contains(1) || pages.contains(400)) {
-            throw new IllegalArgumentException("-1");
-        }
-        return pages;
-    }
-
-    public static List<Integer> numberContinuity(List<Integer> pages) {
-        for (int i = 0; i < pages.size() - 1; i += pages.size() / 2) {
-            if (pages.get(i) + 1 != pages.get(i + 1) && i < pages.size() - 1) {
-                throw new IllegalArgumentException("-1");
-            }
-        }
-        return pages;
-    }
 }
