+++ src/main/java/onboarding/problem1/validator/Validate.java
@@ -1,0 +1,30 @@
+package onboarding.problem1.validator;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Validate {
+    public List<Integer> validate(List<Integer> pobi, List<Integer> crong){
+        List<Integer> page = numPolar(pobi, crong);
+        return numberContinuity(page);
+    }
+
+    public static List<Integer> numPolar(List<Integer> pobi, List<Integer> crong) {
+        List<Integer> pages = new ArrayList<>(pobi);
+        pages.addAll(crong);
+        if (pages.contains(1) || pages.contains(400)) {
+            throw new IllegalArgumentException("-1");
+        }
+        return pages;
+    }
+
+    public static List<Integer> numberContinuity(List<Integer> pages) {
+        for (int i = 0; i < pages.size() - 1; i += pages.size() / 2) {
+            if (pages.get(i) + 1 != pages.get(i + 1) && i < pages.size() - 1) {
+                throw new IllegalArgumentException("-1");
+            }
+        }
+        return pages;
+    }
+
+}
